Вынесены все ручки по jobs в отдельный flask blueprint
